+++ spec/policy/3scale_batcher/3scale_batcher_spec.lua
@@ -29,14 +29,13 @@ describe('3scale batcher policy', function()
   end)
 
   describe('.rewrite', function()
-    it('sets flags to avoid calling backend in the APIcast policy', function()
+    it('sets flag to avoid calling backend in the APIcast policy', function()
       local context = {}
       local batcher_policy = ThreescaleBatcher.new({})
 
       batcher_policy:rewrite(context)
 
       assert.is_true(context.skip_apicast_access)
-      assert.is_true(context.skip_apicast_post_action)
     end)
   end)